Fix native fax linking and automate macOS TIFF runtime dependencies.
Restore stable SpanDSP symbol compatibility for UDPTL/fax builds and add automatic syncing of required libtiff transitive dylibs so parity checks and local builds stay reliable on macOS.

Made-with: Cursor

+++ src-tauri/src/packet_capture/pcap_reader.rs
@@ -238,7 +238,7 @@ impl MmapPcapReader {
 
         let pcap_header = pcap::PacketHeader {
             ts: libc::timeval {
-                tv_sec: header.ts_sec as i32,
+                tv_sec: header.ts_sec as i64,
                 tv_usec: ts_usec as i32,
             },
             caplen: header.caplen,
